+++ packages/eas-cli/src/graphql/mutations/PublishMutation.ts
@@ -1,7 +1,12 @@
 import gql from 'graphql-tag';
 
 import { graphqlClient, withErrorHandlingAsync } from '../client';
-import { PublishUpdateGroupInput, UpdatePublishMutation } from '../generated';
+import {
+  CodeSigningInfoInput,
+  PublishUpdateGroupInput,
+  SetCodeSigningInfoMutation,
+  UpdatePublishMutation,
+} from '../generated';
 
 export const PublishMutation = {
   async getUploadURLsAsync(contentTypes: string[]): Promise<{ specifications: string[] }> {
@@ -43,6 +48,7 @@ export const PublishMutation = {
                   group
                   runtimeVersion
                   platform
+                  manifestPermalink
                 }
               }
             }
@@ -52,5 +58,38 @@ export const PublishMutation = {
         .toPromise()
     );
     return data.updateBranch.publishUpdateGroups;
+  },
+
+  async setCodeSigningInfoAsync(
+    updateId: string,
+    codeSigningInfo: CodeSigningInfoInput
+  ): Promise<SetCodeSigningInfoMutation['update']['setCodeSigningInfo']> {
+    const data = await withErrorHandlingAsync(
+      graphqlClient
+        .mutation<SetCodeSigningInfoMutation>(
+          gql`
+            mutation SetCodeSigningInfoMutation(
+              $updateId: ID!
+              $codeSigningInfo: CodeSigningInfoInput!
+            ) {
+              update {
+                setCodeSigningInfo(updateId: $updateId, codeSigningInfo: $codeSigningInfo) {
+                  id
+                  group
+                  awaitingCodeSigningInfo
+                  codeSigningInfo {
+                    keyid
+                    alg
+                    sig
+                  }
+                }
+              }
+            }
+          `,
+          { updateId, codeSigningInfo }
+        )
+        .toPromise()
+    );
+    return data.update.setCodeSigningInfo;
   },
 };